Practice 09/requirements.txt
pygame>=2.5.0
mutagen>=1.47.0
python-vlc>=3.0.0
